+++ CHANGELOG.md
@@ -12,13 +12,19 @@
 * Add RTL/Jest testing for `AnonymizingTypeSelectContainer` component in `src/settings/components/AnonymizingTypeSelect`. Refs UICIRC-654.
 * Add id for Pane component. Refs UICIRC-756.
 * Fix problem with module launch. Refs UICIRC-768.
+* Cover LoansSection component by RTL/jest tests. Refs UICIRC-613.
+* Add RTL/Jest testing for `RequestPolicyForm` component in `src/settings/RequestPolicy`. Refs UICIRC-647.
 * Cover FixedDueDateScheduleForm component by RTL/jest tests. Refs UICIRC-609.
+
+## [7.0.2](https://github.com/folio-org/ui-circulation/tree/v7.0.2) (2022-04-06)
+[Full Changelog](https://github.com/folio-org/ui-circulation/compare/v7.0.1...v7.0.2)
+
+* Lost Item Policy validation error with aging only recalls to lost. Refs UICIRC-772.
 
 ## [7.0.1](https://github.com/folio-org/ui-circulation/tree/v7.0.1) (2022-03-31)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v7.0.0...v7.0.1)
 
 * Add the possibility to select only active templates. Refs UICIRC-771.
-* Cover LoansSection component by RTL/jest tests. Refs UICIRC-613.
 
 ## [7.0.0](https://github.com/folio-org/ui-circulation/tree/v7.0.0) (2022-02-24)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v6.0.1...v7.0.0)